refactor: 컴포넌트, 커스텀 훅, 상수 파일로 분리
