+++ src/components/playlistForm/PlaylistForm.tsx
@@ -1,131 +1,44 @@
-import { useEffect, useRef, useState } from 'react';
-
 import { css } from '@emotion/react';
 import { RiCloseLine } from 'react-icons/ri';
 
 import Button from '@/components/common/buttons/Button';
 import SelectBox from '@/components/common/SelectBox';
 import ToggleSwitch from '@/components/common/ToggleSwitch';
-import { Video } from '@/mock/data';
-import { useToggleStore } from '@/store/useToggleStore';
+import { CATEGORY_OPTIONS, INITIAL_FORM_STATE } from '@/constants/playlist';
+import { usePlaylistForm } from '@/hooks/usePlaylistForm';
 import theme from '@/styles/theme';
 import { PlaylistFormDataModel } from '@/types/playlist';
 
-const options = [
-  { value: 'lecture', label: '강의' },
-  { value: 'animal', label: '동물' },
-  { value: 'life', label: '라이프' },
-  { value: 'travel', label: '여행' },
-  { value: 'movie', label: '영화' },
-  { value: 'entertainment', label: '엔터테인먼트' },
-  { value: 'music', label: '음악' },
-  { value: 'kids', label: '키즈' },
-];
-
 interface PlaylistFormProps {
-  initialTitle?: string;
-  initialDescription?: string;
-  initialCategory?: string;
-  initialCreatedAt?: string;
-  initialUpdatedAt?: string;
-  initialLikeCount?: number;
-  initialForkCount?: number;
-  initialCommentCount?: number;
-  initialVideoCount?: number;
-  initialThumbnailUrl?: string;
-  initialIsPublic?: boolean;
-  initialVideos?: Video[];
+  initialData?: Partial<PlaylistFormDataModel>;
   onSubmit: (data: PlaylistFormDataModel) => void;
 }
-const PlaylistForm: React.FC<PlaylistFormProps> = ({
-  initialTitle = '',
-  initialDescription = '',
-  initialCategory = '',
-  initialCreatedAt = '',
-  initialUpdatedAt = '',
-  initialLikeCount = 0,
-  initialForkCount = 0,
-  initialCommentCount = 0,
-  initialVideoCount = 0,
-  initialThumbnailUrl = '',
-  initialIsPublic = false,
-  initialVideos = [],
-  onSubmit,
-}: PlaylistFormProps) => {
-  const isToggled = useToggleStore((state) => state.isToggled); // isPublic
-  const toggle = useToggleStore((state) => state.toggle);
-  const [titleValue, setTitleValue] = useState(initialTitle);
-  const [categoryValue, setCategoryValue] = useState<string | undefined>(undefined);
-  const [previewImage, setPreviewImage] = useState<string | null>(null);
-  const titleInputRef = useRef<HTMLInputElement>(null);
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const textAreaRef = useRef<HTMLTextAreaElement>(null);
+const PlaylistForm: React.FC<PlaylistFormProps> = ({ initialData = {}, onSubmit }) => {
+  const {
+    titleValue,
+    categoryValue,
+    previewImage,
+    isToggled,
+    titleInputRef,
+    fileInputRef,
+    textAreaRef,
+    handleCategoryChange,
+    handleFileChange,
+    handleFileButtonClick,
+    handleRemoveImage,
+    handleTitleChange,
+    handleSubmitForm,
+    toggle,
+  } = usePlaylistForm({ ...INITIAL_FORM_STATE, ...initialData, videos: [] });
 
-  useEffect(() => {
-    titleInputRef.current?.focus();
-  }, []);
-
-  useEffect(() => {
-    if (textAreaRef.current) {
-      textAreaRef.current.value = '';
-    }
-  }, []);
-  const handleCategoryChange = (value: string) => {
-    setCategoryValue(value);
-    // 여기서 서버로 데이터 전송
-    // 예시) sendDataToServer(value);
-  };
-
-  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    const reader = new FileReader();
-    reader.onload = () => {
-      // 미리보기 이미지 보여주기
-      setPreviewImage(reader.result as string);
-    };
-    reader.readAsDataURL(file);
-  };
-
-  const handleFileButtonClick = () => {
-    fileInputRef.current?.click();
-  };
-  const handleRemoveImage = () => {
-    setPreviewImage(null);
-    if (fileInputRef.current) {
-      fileInputRef.current.value = ''; // 파일입력 초기화
-    }
-  };
-  const handleTitleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setTitleValue(event.target.value);
-  };
-  const handleSubmitForm = (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    const formData = new FormData(event.target as HTMLFormElement);
-    const { title, category } = Object.fromEntries(formData.entries()) as {
-      title: string;
-      category: string;
-    };
-    const data: PlaylistFormDataModel = {
-      title,
-      description: textAreaRef.current?.value || '',
-      category,
-      createdAt: new Date().toISOString(),
-      updatedAt: new Date().toISOString(),
-      likeCount: 0,
-      forkCount: 0,
-      commentCount: 0,
-      videoCount: 0,
-      thumbnailUrl: previewImage ?? '',
-      isPublic: isToggled,
-      videos: [],
-    };
-    console.log(data);
+  const onFormSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+    const data = handleSubmitForm(event);
     onSubmit(data);
   };
+
   return (
     <div css={containerStyle}>
-      <form onSubmit={handleSubmitForm}>
+      <form onSubmit={onFormSubmit}>
         <input
           type='text'
           name='title'
@@ -133,6 +46,7 @@ const PlaylistForm: React.FC<PlaylistFormProps> = ({
           placeholder='새로운 재생목록 이름 입력하세요 (최대 50자)'
           value={titleValue}
           onChange={handleTitleChange}
+          ref={titleInputRef}
         />
         <div css={divContainerStyle}>
           <span>썸네일(대표이미지)</span>
@@ -170,7 +84,7 @@ const PlaylistForm: React.FC<PlaylistFormProps> = ({
         <div css={divContainerStyle}>
           <span>카테고리</span>
           <SelectBox
-            items={options}
+            items={CATEGORY_OPTIONS}
             onChange={handleCategoryChange}
             value={categoryValue}
             name='category'
